Added definition for BinaryHeap class in BHeap.h
